[feat:#312]: refactor tag value trie read/write
