+++ tsdb/tblstore/invertedindex/series_trie_tree.go
@@ -280,6 +280,7 @@ func (block *trieTreeBlock) FindOffsetsByIn(values []string) (offsets []int) {
 }
 
 func (block *trieTreeBlock) FindOffsetsByLike(value string) (offsets []int) {
+	//FIXME stone1100
 	switch value {
 	case "":
 		return nil
